Move filler quirks handling from the TreeView to the filler utils.
